+++ react/src/dashboard/props/models/Prevalidation.tsx
@@ -1,4 +1,5 @@
 import PrevalidationRecord from './PrevalidationRecord';
+import Fund from './Fund';
 
 export default interface Prevalidation {
     id: number;
@@ -8,6 +9,7 @@ export default interface Prevalidation {
     identity_address: string;
     records_hash: string;
     state: string;
+    fund?: Fund;
     uid?: string;
     uid_hash?: string;
 }
