+++ react/src/dashboard/services/FundService.ts
@@ -40,10 +40,6 @@ export class FundService<T = Fund> {
         return this.apiRequest.get(`${this.prefix}/${company_id}/funds`, data);
     }
 
-    public listPublic(data: object = {}): Promise<ResponseSimple<{ data: Array<T> }>> {
-        return this.apiRequest.get(`${this.prefix_public}`, data);
-    }
-
     public read(company_id: number, fund_id: number, data: object = {}): Promise<ApiResponseSingle<T>> {
         return this.apiRequest.get(`${this.prefix}/${company_id}/funds/${fund_id}`, data);
     }
