+++ react/src/dashboard/services/PrevalidationService.ts
@@ -15,26 +15,31 @@ export class PrevalidationService<T = Prevalidation> {
      *
      * @param data
      */
-    public prefix = '/platform/prevalidations';
+    public prefix = '/platform/organizations';
 
-    public list(filters: object = {}): Promise<ApiResponse<T>> {
-        return this.apiRequest.get(`${this.prefix}`, filters);
+    /**
+     * @param organization_id
+     * @param filters
+     */
+    public list(organization_id: number, filters: object = {}): Promise<ApiResponse<T>> {
+        return this.apiRequest.get(`${this.prefix}/${organization_id}/prevalidations`, filters);
     }
 
-    public submit(data: object = {}, fund_id: number = null): Promise<ApiResponseSingle<T>> {
-        return this.apiRequest.post(`${this.prefix}`, {
+    public store(organization_id: number, data: object = {}, fund_id: number = null): Promise<ApiResponseSingle<T>> {
+        return this.apiRequest.post(`${this.prefix}/${organization_id}/prevalidations`, {
             data: data,
             fund_id: fund_id,
         });
     }
 
-    public submitCollection(
+    public storeBatch(
+        organization_id: number,
         data: Array<{ [key: string]: string }>,
         fund_id: number = null,
         overwrite: Array<string> = [],
         file?: object,
     ): Promise<T> {
-        return this.apiRequest.post(`${this.prefix}/collection`, {
+        return this.apiRequest.post(`${this.prefix}/${organization_id}/prevalidations/collection`, {
             data: data,
             fund_id: fund_id,
             overwrite: overwrite,
@@ -43,6 +48,7 @@ export class PrevalidationService<T = Prevalidation> {
     }
 
     public submitCollectionCheck(
+        organization_id: number,
         data: Array<string>,
         fund_id: number = null,
         overwrite: Array<string> = [],
@@ -52,33 +58,33 @@ export class PrevalidationService<T = Prevalidation> {
             db: Array<{ uid_hash: string; records_hash: string }>;
         };
     }> {
-        return this.apiRequest.post(`${this.prefix}/collection/hash`, {
+        return this.apiRequest.post(`${this.prefix}/${organization_id}/prevalidations/collection/hash`, {
             data: data,
             fund_id: fund_id,
             overwrite: overwrite,
         });
     }
 
-    public destroy(code: string): Promise<ApiResponseSingle<T>> {
-        return this.apiRequest.delete(`${this.prefix}/${code}`);
+    public destroy(organization_id: number, code: string): Promise<ApiResponseSingle<T>> {
+        return this.apiRequest.delete(`${this.prefix}/${organization_id}/prevalidations/${code}`);
     }
 
-    public export(filters: object = {}): Promise<ResponseSimple<ArrayBuffer>> {
-        return this.apiRequest.get(`${this.prefix}/export`, filters, {
+    public export(organization_id: number, filters: object = {}): Promise<ResponseSimple<ArrayBuffer>> {
+        return this.apiRequest.get(`${this.prefix}/${organization_id}/prevalidations/export`, filters, {
             responseType: 'arraybuffer',
         });
     }
 
     public getColumns(headers: Array<string>, typesByKey: object): Array<ConfigurableTableColumn> {
-        const list = ['code', 'employee', ...headers, 'active', 'exported'];
+        const list = ['code', 'fund', 'employee', ...headers, 'active', 'exported'];
 
         return list.map((key) => ({
             key,
-            label: typesByKey[key] || `prevalidated_table.labels.${key}`,
+            label: typesByKey?.[key] || `prevalidated_table.labels.${key}`,
             tooltip: {
                 key: key,
-                title: typesByKey[key] || `prevalidated_table.labels.${key}`,
-                description: typesByKey[key] || `prevalidated_table.tooltips.${key}`,
+                title: typesByKey?.[key] || `prevalidated_table.labels.${key}`,
+                description: typesByKey?.[key] || `prevalidated_table.tooltips.${key}`,
             },
         }));
     }
